% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adis2csv2.R
\name{adis2csv2}
\alias{adis2csv2}
\title{Directly writes the content of an ADIS data file to multiple .csv files.}
\usage{
adis2csv2(file, output_path = NULL, entities = NULL, ...)
}
\arguments{
\item{file}{Character of length 1. Full path to the ADIS file to be read, expected in UNIX-style ('/').}

\item{output_path}{Character of length 1. Where shall the .csv files be saved ? (expected in UNIX-style ('/')).
If \code{NULL} (default), the working directory is used.}

\item{entities}{Character vector. Which entities should be exported? Is recycled for all headers.
If \code{NULL} (default), all.}

\item{...}{Further arguments passed to \link{read_adis}.}
}
\description{
Do you really want to use another software than \code{R} to analyse your data???
- \code{adis2csv2()} helps you with this sad task...
It uses \link{read_adis} to read ADIS files to R and directly writes all entities as .csv files in subdirectories of
\code{output_path} for each header (.csv files are written by \link{write.csv2}).
}
\note{
Despite using \link{write.csv2}, the decimal points in the .csv files will be '.' because all items
 are imported as character vectors.
}

