% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_adis.R
\name{read_adis}
\alias{read_adis}
\title{Reading ADIS data files}
\usage{
read_adis(datei, zuNA = c(" ", "?"), entitiesmehrfach = TRUE,
  leerloeschen = TRUE, kommentare = TRUE)
}
\arguments{
\item{datei}{Character of length 1. Full path to the ADIS file to be read, expected in UNIX-style ('/').}

\item{zuNA}{Character vector. Which characters should be interpreted as NA?}

\item{entitiesmehrfach}{Logical of length 1. Should entities that are duplicated under one header be read multiple times?
(Default is \code{TRUE}) If \code{FALSE} only the first is read with a warning.}

\item{leerloeschen}{Logical of length 1. Should unnecessary spaces be omitted? (Default is \code{TRUE}).}

\item{kommentare}{Logical of length 1. (Default is \code{TRUE}) If \code{TRUE} comment lines of the ADIS file
 are provided together as first element of the output, otherwise they are simply ignored.}
}
\value{
\code{read_adis()} returns a list of the headers of the ADIS file with dataframes for each entity.
 The item names within the entities are returned as column names in the dataframes.
 All item contents are returned as character vectors (columns in the dataframes).
}
\description{
The workhorse function of \pkg{RAdis}.
}

